+++ wyoming/intent.py
@@ -9,6 +9,8 @@ DOMAIN = "intent"
 _RECOGNIZE_TYPE = "recognize"
 _INTENT_TYPE = "intent"
 _NOT_RECOGNIZED_TYPE = "not-recognized"
+_INTENTS_START_TYPE = "intents-start"
+_INTENTS_STOP_TYPE = "intents-stop"
 
 
 @dataclass
@@ -41,7 +43,6 @@ class Recognize(Eventable):
 
     @staticmethod
     def from_event(event: Event) -> "Recognize":
-        assert event.data is not None
         return Recognize(text=event.data["text"], context=event.data.get("context"))
 
 
@@ -95,7 +96,6 @@ class Intent(Eventable):
 
     @staticmethod
     def from_event(event: Event) -> "Intent":
-        assert event.data is not None
         return Intent.from_dict(event.data)
 
     def to_rhasspy(self) -> Dict[str, Any]:
@@ -136,7 +136,72 @@ class NotRecognized(Eventable):
 
     @staticmethod
     def from_event(event: Event) -> "NotRecognized":
-        assert event.data is not None
         return NotRecognized(
             text=event.data.get("text"), context=event.data.get("context")
         )
+
+
+@dataclass
+class IntentsStart(Eventable):
+    """Start of one or more intents.
+
+    Event flow:
+    intents-start
+    intent
+    [intent]...
+    intents-stop
+    """
+
+    context: Optional[Dict[str, Any]] = None
+    """Context for next interaction."""
+
+    @staticmethod
+    def is_type(event_type: str) -> bool:
+        return event_type == _INTENTS_START_TYPE
+
+    def event(self) -> Event:
+        data: Dict[str, Any] = {}
+        if self.context is not None:
+            data["context"] = self.context
+
+        return Event(type=_INTENTS_START_TYPE, data=data)
+
+    @staticmethod
+    def from_event(event: Event) -> "IntentsStart":
+        if not event.data:
+            return IntentsStart()
+
+        return IntentsStart(context=event.data.get("context"))
+
+
+@dataclass
+class IntentsStop(Eventable):
+    """End of intent series.
+
+    Event flow:
+    intents-start
+    intent
+    [intent]...
+    intents-stop
+    """
+
+    context: Optional[Dict[str, Any]] = None
+    """Context for next interaction."""
+
+    @staticmethod
+    def is_type(event_type: str) -> bool:
+        return event_type == _INTENTS_STOP_TYPE
+
+    def event(self) -> Event:
+        data: Dict[str, Any] = {}
+        if self.context is not None:
+            data["context"] = self.context
+
+        return Event(type=_INTENTS_STOP_TYPE, data=data)
+
+    @staticmethod
+    def from_event(event: Event) -> "IntentsStop":
+        if not event.data:
+            return IntentsStop()
+
+        return IntentsStop(context=event.data.get("context"))